+++ core/src/stored/label.cc
@@ -625,22 +625,8 @@ bool WriteSessionLabel(DeviceControlRecord* dcr, int label)
 
   rec = new_record();
   Dmsg1(130, "session_label record=%x\n", rec);
-  switch (label) {
-    case SOS_LABEL:
-      SetStartVolPosition(dcr);
-      break;
-    case EOS_LABEL:
-      if (dev->IsTape()) {
-        dcr->EndBlock = dev->EndBlock;
-        dcr->EndFile = dev->EndFile;
-      } else {
-        dcr->EndBlock = (uint32_t)dev->file_addr;
-        dcr->EndFile = (uint32_t)(dev->file_addr >> 32);
-      }
-      break;
-    default:
-      Jmsg1(jcr, M_ABORT, 0, _("Bad Volume session label = %d\n"), label);
-      break;
+  if (label != SOS_LABEL && label != EOS_LABEL) {
+    Jmsg1(jcr, M_ABORT, 0, _("Bad Volume session label = %d\n"), label);
   }
   CreateSessionLabel(dcr, rec, label);
   rec->FileIndex = label;